requirements.txt
amqp==2.1.1
anyjson==0.3.3
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.1
billiard==3.5.0.2
bs4==0.0.1
celery==4.0.0
Django==1.10.5
django-braces==1.11.0
django-cart==1.0.4
django-celery==3.1.17
django-cors-middleware==1.3.1
django-datatable==0.2.1
django-datatables-view==1.14.0
django-easy-pdf==0.1.0
django-filter==1.0.1
django-mailchimp-v1.3==1.4.0
django-money==0.10.1
django-moneyfield==0.2.1
django-oauth-toolkit==0.12.0
django-oauth2-provider==0.2.6.1
django-paypal==0.3.3
django-render-block==0.5
django-templated-email==2.1
djangorestframework==3.3.2
docopt==0.4.0
gunicorn==19.7.0
html5lib==1.0b8
kombu==4.0.0
mailchimp==2.0.9
mailchimp3==2.0.7
Markdown==2.6.8
MySQL-python==1.2.5
mysqlclient==1.3.9
oauthlib==2.0.1
olefile==0.44
passlib==1.7.1
pdfkit==0.6.1
Pillow==4.0.0
py-moneyed==0.6.0
pycountry==17.1.8
pycrypto==2.6.1
Pygments==2.2.0
PyPDF2==1.26.0
python-dateutil==1.5
python-memcached==1.58
pytz==2016.7
reportlab==3.3.0
requests==2.12.5
shortuuid==0.5.0
six==1.10.0
socketIO-client==0.7.2
vine==1.1.3
webencodings==0.5
websocket-client==0.40.0
whitenoise==3.3.0
xhtml2pdf==0.0.6
